Attempt to eliminate burns on Windows due to erroniously assigning size_t to int
svn:sync-xref-src-repo-rev: r129236

+++ src/VBox/Frontends/VBoxManage/VBoxManageDisk.cpp
@@ -1,5 +1,5 @@
 #include <signal.h>
-/* $Id: VBoxManageDisk.cpp 77606 2019-03-08 02:55:07Z noreply@oracle.com $ */
+/* $Id: VBoxManageDisk.cpp 77608 2019-03-08 04:01:48Z noreply@oracle.com $ */
 /** @file
  * VBoxManage - The disk/medium related commands.
  */
@@ -324,7 +324,7 @@ RTEXITCODE handleCreateMedium(HandlerArg *a)
                 PMEDIUMPROPERTY pNewProp = (PMEDIUMPROPERTY)RTMemAlloc(sizeof(MEDIUMPROPERTY));
                 if (!pNewProp)
                     return errorArgument("Can't allocate memory for property '%s'", ValueUnion.psz);
-                int cbKvp = RTStrNLen(ValueUnion.psz, PROP_MAXLEN);
+                size_t cbKvp = RTStrNLen(ValueUnion.psz, PROP_MAXLEN);
                 char *cp;
                 for (cp = (char *)ValueUnion.psz; *cp != '=' && cp < ValueUnion.psz + cbKvp; cp++)
                     continue;
